LPD-51750 Split frontend-js-web's global submodule in two parts
We need a preload part that will be loaded from the very begining, right
after AUI is loaded and while the Liferay global object is populated with
the most low level stuff.

The rest will go in the global submodule that will be loaded via ESM and
will re-export stuff from the main submodule through the Liferay global
object.

The intention is to slowly migrate all usages of Liferay global object in
the ES code to proper imports so that we can benefit from optimizations in
the future. That's why the preload stuff is reexported from the Liferay
object through the main submodule (as opposed to the reverse direction for
non preload stuff).