LPD-52709 Remove LPD-11848 feature flag and disable LanguageFilter for internal ESM files
We need to use language labels modules for caching to be feasible so we will split .js modules in
two parts: code itself and labels in a different file, thus we don't need the LanguageFilter to kick
in any more.

This will make the use of ETags for .js files not necessary any more.

Note however that legacy /combo URLs are still routed through the filter. In any case this should
only be used for YUI code, not for ESM.