LPD-52709 Add the frontend resources filter and the language modules handler
This commit adds the filter that will manage all type of frontend resources
and implements the handling of labels modules since we removed the language
servlet in the previous commit.

The AbstractRequestHelper class encapsulates the common logic for hashed vs
non-hashed (virtual) resources caching.

The LanguageRequestHelper class implements the handling of language labels
modules. It contains the logic that is specific to those resources.