LPD-52709 Remove the max-age window for labels modules
After discussing in a PR review we have decided to make the model (and
code) simpler at the expense of guaranteeing label updates in at most
2*max-age seconds instead of 1*max-age seconds.

This is because we favor simplicity and don't see too much value in halving
the window when there are a lot of different scenarios that may lead to
a browser using the "wrong" labels not only during expiration of the cache
but when loading the labels in time.

So for now we will keep things simple and we'll see what happens in the
future.

Additionally, having a rolling window may be bad because it would create
bursts of updates in the server where everybody re-requests all labels
files at the same time instead of sparsing the requests over time.