WIP "fix" flaky tests
I'm adding delays to failing tests to demonstrate that the tests are
inherently flaky in nature and must be reworked.

The PR that makes tests fails modifies the way in which JS files are
served. These introduce timing variations in the page loading process which
in turn make flaky tests pop out even if they were "stable" before (due to
the state of things previous to the PR).