LPD-68814 Lower log level
The ERROR level was breaking some CI tests because PortletRenderUtil is
logging the message when it cannot find a CSS file. We used error because
we assumed all CSS files requested by that class should exist, but in the
case of tests, they don’t.

Now we lower the log message to DEBUG because the tokenized feature is only
used for legacy code and if a CSS file that has tokens cannot be found by
PortletRenderUtil and it ends up causing a visible error, it should be a
fairly straightforward process to diagnose it, activating the traces if
necessary.

(see https://liferay.atlassian.net/browse/LPD-68304?focusedCommentId=30544639)