LPD-67267 TypeScript integration
This is a way to tell TypeScript where to get the types for CKEditor split
packages.

Another option would be using bare identifiers in source code (for example:
`@ckeditor/ckeditor5-core` instead of
`@ckeditor/ckeditor5-core/dist/index.js`, then configuring aliases to
rewrite imports from `@ckeditor/ckeditor5-core` to
`@ckeditor/ckeditor5-core/dist/index.js`.

It seems to me that the `types.d.ts` option is easier to maintain and also
clearer since what you see in code is what you get at runtime.

In the end, it's a TypeScript problem that has nothing to do with code and
bundling so it wouldn't be fair to affect those two.