LPD-52709 Implement translation of portlet JS files to avoid breaking changes
Since portlet JS files are not intended to be ESM format and outside of
that context we cannot use the new translation architecture based on all.js
files, I'm putting back again portlet JS files translation. It essentially
works the same as tokenized CSS files.

Note that translated JS files will not be infinitely cacheable but
time-based, for obvious reasons.