LPD-70652 Suppress @typescript-eslint/unused-vars false positives
Sometimes @typescript-eslint/unused-vars emits false positives due to the
way TypeScript parsing for eslint works.

The canonical way to fix those false positives is updating to the latest
version but even in that case the false positives may still appear.

Additionally, updating to the latest version means we must update (a lot)
@liferay/eslint-plugin, which may require a lot of work and still not fix
the issue. So I'm simply ignoring these false positives for now.