LPD-75157 we cannot check for themeDisplay.isShowControlMenu() inside ProductNavigationProductMenuHelper. That's because it is called by ProductMenuProductNavigationControlMenuEntry#isShow, which is called internally in ProductNavigationControlMenuManagerImpl#isShowControlMenu.
When a page is going to load, 'ProductNavigationControlMenuManagerImpl#isShowControlMenu' runs,
which goes over all entries that the user can view. One of the entries is
'ProductMenuProductNavigationControlMenuEntry', which calls
'ProductNavigationProductMenuHelperImpl#isShowProductMenu'. That is, if I
check for 'themeDisplay.isShowControlMenu()' there, it won't have a value
yet, because I am still calculating it. Therefore, since there are only
three usages of ProductNavigationProductMenuHelper#isShowProductMenu, I
can just check for 'themeDisplay.isShowControlMenu()' on those usages,
because they happen after
'ProductNavigationControlMenuManagerImpl#isShowControlMenu'.